run `npx svelte-migrate routes`
