+++ src/routes/+page.svelte
@@ -1,7 +1,3 @@
-<script context="module" lang="ts">
-	export const prerender = true;
-</script>
-
 <script>
 	$: innerWidth = 0;
 </script>
