+++ src/routes/about/+page.js
@@ -1,0 +1,6 @@
+export async function load({ fetch }) {
+	const articles = await fetch('/about.json').then((r) => r.json());
+	return {
+		post: articles.index
+	};
+}
