+++ src/routes/about/+page.svelte
@@ -1,15 +1,6 @@
-<script context="module">
-	export async function load({ fetch }) {
-		const articles = await fetch('/about.json').then((r) => r.json());
-		return {
-			props: {
-				post: articles.index
-			}
-		};
-	}
-</script>
-
 <script>
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
 	export let post;
 	import TapedPolaroid from '$lib/TapedPolaroid.svelte';
 	import Article from '../../pages/Article.svelte';
