+++ src/routes/posts/+page.ts
@@ -1,0 +1,1 @@
+export const prerender = true;
